test verify-commit/tag to exit unsuccessfully
git verify-commit and verify-tag should exit unsuccessfully when
processing a signature by a gpg key with trust level set to 'never'.

This commit introduce verify checks with trust-model set to direct in
gpg.conf (to force trust level of the second key in the keychain to
never). In these tests, 'git verify-tag/verify-commit eighth-signed-alt'
must exit unsuccessfully and includes 'We do NOT trust this key!' on the
stderr (gpg output).

Helped-by: Karel Koci <karel.koci@nic.cz>